fix(tof): carry monotonic time in 64 bits, so an alarm cannot switch itself off [AMRSW-2322]
source_status() decided HEALTHY with a SIGNED 32-bit comparison, and it had to: the reference
for a source that had never published was `startup_time + STARTUP_GRACE_MS`, a point in the
FUTURE during the grace period, so an unsigned subtraction would have read as an enormous
elapsed time and reported every warming-up sensor as stale.

The cost of that arrangement was paid 24.8 days later. Once `now - reference` passed 2^31 the
cast read negative, a long-dead source was classified HEALTHY, and poll() then

  - emitted SOURCE_RECOVERED for a sensor that had never come back,
  - cleared alarm_active, and
  - silenced report_persistent_state(), which gates on it.

So the failure was not a delayed alarm. It was an alarm that switched itself off and left an
actively false "recovered" line in rosout. Separately, monotonic_ms() truncated steady_clock
to 32 bits, putting a second wrap at 49.7 days under the same code.

Fixed by removing the wrap rather than arranging arithmetic around it. Every timestamp in the
assembler, the receiver, the tracker and slot state, the reported ages and the report throttle
is now uint64_t, and the comparison is restructured so no reference is ever in the future:
the startup grace is added to the BUDGET, not to the reference, which leaves every subtraction
unsigned.

  never published : elapsed(startup_time, now) <= STARTUP_GRACE_MS + SOURCE_STALE_MS
  has published   : elapsed(last_publish, now) <= SOURCE_STALE_MS

Semantically identical to the signed form for all reachable times, and correct beyond it. The
two time-comparison conventions that used to sit in one function are now one.

Four tests, each confirmed by mutation -- the test fails when the production rule is removed:

  - NoFalseRecoveryBeyondTheSignedRange / StillStaleBeyondThe32BitWrap fail if the signed
    32-bit comparison returns. The 2^32 case lands deliberately only 500 ms past the wrap:
    truncated to 32 bits that is inside SOURCE_STALE_MS of the truncated reference, which is
    what makes a broken implementation report a freshly published source. Landing further past
    the wrap would have passed on the bug.
  - OnlyACompletedGridClearsTheAlarm fails if any arriving frame clears alarm_active. It checks
    the flag before the next poll(), because a poll would re-raise the alarm on a still-stale
    source and hide exactly that mutation.
  - StartupGraceBoundaryIsUnchanged fails if the grace is dropped from the budget, which is the
    one place the restructure could have shifted the boundary by a millisecond.

+++ src/receiver.cpp
@@ -65,19 +65,19 @@ public:
   {
   }
 
-  void enable_tof_can(uint32_t data_id, uint32_t health_id, uint32_t now_ms)
+  void enable_tof_can(uint32_t data_id, uint32_t health_id, uint64_t now_ms)
   {
     tof_data_can_id = data_id;
     tof_health_can_id = health_id;
     tof.configure_can(data_id, health_id, now_ms);
   }
 
-  void poll_tof(uint32_t now_ms)
+  void poll_tof(uint64_t now_ms)
   {
     tof.poll(now_ms);
   }
 
-  void handle_can(const can_frame& frame, uint32_t now_ms)
+  void handle_can(const can_frame& frame, uint64_t now_ms)
   {
     if (tof_data_can_id && (frame.can_id == *tof_data_can_id || frame.can_id == *tof_health_can_id))
     {
@@ -141,10 +141,13 @@ private:
   receiver_tof tof;
 };
 
-uint32_t monotonic_ms()
+// 64 bits, deliberately. Truncating steady_clock to 32 bits wrapped every 49.7 days and put
+// the whole ToF watchdog on a clock that could run backwards; every consumer of this value now
+// takes uint64_t so the wrap is gone rather than compensated for.
+uint64_t monotonic_ms()
 {
   using namespace std::chrono;
-  return static_cast<uint32_t>(duration_cast<milliseconds>(steady_clock::now().time_since_epoch()).count());
+  return static_cast<uint64_t>(duration_cast<milliseconds>(steady_clock::now().time_since_epoch()).count());
 }
 
 tof_transport resolve_tof_transport(ros::NodeHandle& pn)
@@ -292,8 +295,8 @@ int main(int argc, char* argv[])
     }
   }
 
-  uint32_t last_tof_poll_ms = monotonic_ms();
-  constexpr uint32_t tof_poll_interval_ms = 100;
+  uint64_t last_tof_poll_ms = monotonic_ms();
+  constexpr uint64_t tof_poll_interval_ms = 100;
 
   // Start I/O threads
   std::atomic<bool> running{ true };
@@ -328,7 +331,7 @@ int main(int argc, char* argv[])
     // The watchdog has to run whether or not frames are arriving; a silent source is
     // precisely the case consume() can never see. Throttled because the surrounding loop
     // does not rate limit itself.
-    uint32_t const now = monotonic_ms();
+    uint64_t const now = monotonic_ms();
     if (now - last_tof_poll_ms >= tof_poll_interval_ms)
     {
       last_tof_poll_ms = now;
